rename cloud-init module files to test_, merge two duplicate testfiles
