+++ tests/unittests/config/test_cc_apk_configure.py
