+++ tests/unittests/config/test_cc_bootcmd.py
