+++ tests/unittests/config/test_cc_disable_ec2_metadata.py
