+++ tests/unittests/config/test_cc_disk_setup.py
