+++ tests/unittests/config/test_cc_install_hotplug.py
