+++ tests/unittests/config/test_cc_locale.py
