+++ tests/unittests/config/test_cc_mcollective.py
