+++ tests/unittests/config/test_cc_puppet.py
