+++ tests/unittests/config/test_cc_rsyslog.py
