+++ tests/unittests/config/test_cc_set_hostname.py
