+++ tests/unittests/config/test_cc_ubuntu_advantage.py
