+++ tests/unittests/config/test_cc_zypper_add_repo.py
